+++ packages/just-func/ts/index.ts
@@ -1,2 +1,1 @@
-export * from './Just'
-export * from './toJustValue'
+export * from '@just-func/types'