+++ packages/types/jest.config.js
@@ -1,0 +1,27 @@
+module.exports = {
+  'collectCoverageFrom': [
+    '<rootDir>/ts/**/*.[jt]s',
+    '!<rootDir>/ts/bin.[jt]s',
+  ],
+  'reporters': [
+    'default',
+    'jest-progress-tracker',
+  ],
+  'roots': [
+    '<rootDir>/ts',
+  ],
+  'testEnvironment': 'node',
+  'testMatch': ['**/?(*.)+(spec|test|integrate|accept|system|unit).[jt]s?(x)'],
+  'watchPlugins': [
+    'jest-watch-suspend',
+    'jest-watch-repeat',
+    'jest-watch-typeahead/filename',
+    'jest-watch-typeahead/testname',
+    [
+      'jest-watch-toggle-config', { 'setting': 'verbose' },
+    ],
+    [
+      'jest-watch-toggle-config', { 'setting': 'collectCoverage' },
+    ],
+  ],
+}