+++ packages/just-func/CHANGELOG.md
@@ -1,5 +1,16 @@
 # just-func
 
+## 0.1.1
+
+### Patch Changes
+
+- Move types related code to `@just-func/types`.
+
+  `just-func` export what's in `@just-func/types`
+
+- Updated dependencies
+  - @just-func/types@0.1.0
+
 ## 0.1.0
 
 ### Minor Changes